Rename and restructure into dose/ directory
