+++ tests/unit_tests/test_data_mu_en_coefficients.py
@@ -1,18 +1,20 @@
 from pytest import approx, raises
 
-from openmc.data import mu_en_coefficients
+from openmc.data import mass_energy_absorption_coefficient
+from openmc.data.function import Tabulated1D
 
 
-def test_mu_en_coefficients():
+def test_mass_energy_absorption_coefficient():
     # Spot checks on values from NIST tables
-    energy, mu_en = mu_en_coefficients("air")
-    assert energy[0] == approx(1e3)
-    assert mu_en[0] == approx(3.599e3)
-    assert energy[-1] == approx(2e7)
-    assert mu_en[-1] == approx(1.311e-2)
+    mu_en = mass_energy_absorption_coefficient("air")
+    assert isinstance(mu_en, Tabulated1D)
+    assert mu_en.x[0] == approx(1e3)
+    assert mu_en.y[0] == approx(3.599e3)
+    assert mu_en.x[-1] == approx(2e7)
+    assert mu_en.y[-1] == approx(1.311e-2)
 
-    # Invalid particle/geometry should raise an exception
+    # Invalid material/data_source should raise an exception
     with raises(ValueError):
-        mu_en_coefficients("pasta")
+        mass_energy_absorption_coefficient("pasta")
     with raises(ValueError):
-        mu_en_coefficients("air", data_source="nist000")
+        mass_energy_absorption_coefficient("air", data_source="nist000")
